Moved set_role commands to features
