+++ features/management/admin/__init__.py
@@ -1,1 +1,3 @@
 from . import config_report
+from . import set_role
+from .feature import *
