+++ features/management/admin/set_role/feature.py
@@ -1,6 +1,4 @@
 import logging
-from . import admin
-from . import owner
 
 _logger = logging.getLogger("main")
 
@@ -17,6 +15,4 @@ def is_active():
 def activate():
     global _active
     _active = True
-    _logger.debug("features: Feature has been activated: 'management'")
-    admin.activate()
-    owner.activate()
+    _logger.debug("features: Feature has been activated: 'management/admin/set_role'")
